support docker-machine
docker-machine runs containers in a virtualbox vm instead of using qemu.
Thus, the networking is a bit different because the "host" that exposes
ports is a virtual host.  I solved this by getting the IP address of
that host and binding to that IP address and rewriting the URL for
jupyter notebook.

I also included an untested attempt to open the URL in WSL using
powershell.exe.  The previous implementation was definitely broken, so
no harm done.