Github workflow should test verticalab changes
It /was/ getting the cached verticalab image and before testing
connectivity between verticalab and vertica.  Now it builds a new
verticalab so it will detect new connectivity issues.

Also, I added a verticalab image version and instructions on when to
change it and how to make changes to verticalab such that the main
branch always works.