fix VERTICA_PORT configuration parameter
Setting VSQL_PORT was the wrong thing to do since it then tries to
connect inside the container to the wrong port.  Instead, we need to
keep VERTICA_PORT and VSQL_PORT separate since one is the inside port
and the other is the outside port.
Also removed a temporary kludge with the startup script
Also reduced the redundancy of hardcoded port numbers