move all defaults into conf file
Reduce the redundancy of the default values by moving everything into
the conf file.  When a git-pull causes the conf file to change, we throw
an error that the conf file needs to be manually updated instead of
ignoring the changes of overwriting the configuration.

When a new conf file is created, remove any stopped vertica-demo
containers so that a vertica-start will use the new config.

Fixed a couple instances where the image name didn't start with vertica/

Require a version specified if we are building a new verticalab

Fixed a bad variable name "CONTAINER_NAME" that should have been
"VERTICA_CONTAINER_NAME"

When an open command doesn't exist (linux without X), then change the
text "if this link doesn't open" to just "copy/paste this"

hide some strange output like "is not running" in vertica-start

Fixed shutdown functionality in vertica-start if database is running